he's alive
st-link started working again so i was able to reprogram this version of
code, but DFU is still not working. apparently the bootloader is in ROM,
so it shouldn't be corrupted. weird behavior though: whenever I
connected st-link, the motor runs which might be the reason behind power
failures when using st-link. maybe it's working because the digilent is
not connected to the usb hub now? probably shouldn't be writing this in
a commit message